Small refactoring; removed one layer of unnecessary abstraction.
